* Fixed a small error at game.cpp, thanks Evan. * Now you can use double values for experience, skill, magic rates and experience stages. Issue #2260
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/branches/tryller@5958 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ game.h
@@ -542,7 +542,7 @@ class Game
 		bool npcSpeakToPlayer(Npc* npc, Player* player, const std::string& text, bool publicize);
 
 		bool loadExperienceStages();
-		uint64_t getExperienceStage(uint32_t level);
+		double getExperienceStage(uint32_t level);
 
 		void setServerSaveMessage(int16_t key, bool value) {serverSaveMessage[key] = value;}
 		bool getServerSaveMessage(int16_t key) const {return serverSaveMessage[key];}
